Added Other category, moved Filters and Checkbox to directories & moved title down
