+++ src/components/LocationInput/LocationInput.jsx
@@ -2,28 +2,29 @@ import React from "react";
 
 import PlacesAutocomplete, {
   geocodeByAddress,
-  getLatLng,
+  getLatLng
 } from "react-places-autocomplete";
 import styles from "./LocationInput.module.scss";
 
 class LocationInput extends React.Component {
   state = { address: "" };
 
-  handleChange = (address) => {
+  handleChange = address => {
     this.setState({ address });
   };
 
-  handleSelect = (address) => {
+  handleSelect = address => {
     geocodeByAddress(address)
-      .then((results) => getLatLng(results[0]))
-      .then((latLng) => this.props.getLocationInfo(latLng, address))
-      .catch((error) => console.error("Error", error));
+      .then(results => getLatLng(results[0]))
+      .then(latLng => this.props.getLocationInfo(latLng, address))
+      .catch(error => console.error("Error", error));
   };
 
   render() {
     return (
       <PlacesAutocomplete
         value={this.state.address}
+        searchOptions={{ type: "restaurant" }}
         onChange={this.handleChange}
         onSelect={this.handleSelect}
       >
@@ -32,12 +33,12 @@ class LocationInput extends React.Component {
             <input
               {...getInputProps({
                 placeholder: "Search Places ...",
-                className: "location-search-input",
+                className: "location-search-input"
               })}
             />
             <div className={styles.autocompleteDropdown}>
               {loading && <div>Loading...</div>}
-              {suggestions.map((suggestion) => {
+              {suggestions.map(suggestion => {
                 const className = suggestion.active
                   ? styles.suggestionItemActive
                   : styles.suggestionItem;
@@ -45,7 +46,7 @@ class LocationInput extends React.Component {
                 return (
                   <div
                     {...getSuggestionItemProps(suggestion, {
-                      className,
+                      className
                     })}
                   >
                     <span>{suggestion.description}</span>
